Add files via upload
A coffee shop website is an online platform designed to represent and promote a coffee shop, providing visitors with information about the shop, its offerings, and services. The main purpose of such a website is to engage potential customers and create an inviting digital space that mirrors the ambiance of the physical shop.